Merge changed scopes instead of replacing the Array
There will only ever be one reference to the scope array used, so it
can't be simply replaced when updated. Change to clearing it's current
contents and adding all the new contents while keeping the same array.

Also ensures that the embedded scope is properly kept in the array if
enabled.