Allow JSHint to handle configuration where possible
Use the configuration search only to determine whether or not we should
call the full JSHint run, unless the user has specified a custom name
for the file. If we don't find a file in the default "all parent's"
search, then check the user's home directory and the "disable when no
config" setting to determine whether the lint run should continue.