ci(travis): update to the Xenial based image
Atom's current beta build throws a warning on older versions of the 
`libdbus-1.so.3` library... which the `trusty` based distribution just 
happens to be using. Although this warning doesn't prevent Atom itself 
from working, it causes `stderr` output from the calls to `jshint`, 
which makes it look like it has failed.